arrays: Change i8 to i32

+++ examples/arrays/arrays.zig
@@ -7,11 +7,11 @@ pub fn main() anyerror!void {
 
     // One way to create an array is by using an _array literal_. The compiler
     // can infer its size if you specify it as `_`.
-    const a = [3]i8{ 1, 2, 3 };
-    const b = [_]i8{ 4, 5, 6 };
+    const a = [3]i32{ 1, 2, 3 };
+    const b = [_]i32{ 4, 5, 6 };
 
     // Another way to create an array is by using _anonymous list literals_.
-    const c: [3]i8 = .{ 7, 8, 9 };
+    const c: [3]i32 = .{ 7, 8, 9 };
 
     // All arrays have a `len` field, allowing you to see their size.
     std.debug.print("{}\n", .{c.len});
@@ -19,5 +19,5 @@ pub fn main() anyerror!void {
     // You can concatenate and repeat arrays if they are both compile-time
     // known.
     std.debug.print("{any}\n", .{a ++ b});
-    std.debug.print("{any}\n", .{[_]i8{0} ** 6});
+    std.debug.print("{any}\n", .{[_]i32{0} ** 6});
 }
